HDS-47 clean up logging in/out and resetting state correctly
